Update fltinv man page
The new model uncertainty option is added to the man page.

The detailed description of the annealing iteration reset option
is corrected.

The man page did not have the complete information for the -xy and
-geo options to define coordinate type. These are now included in
the option list and have descriptions.

A brief description of the simulated annealing algorithm is
included, along with corresponding references.