Add program tri_tool
Many deformation applications use triangular meshes, from slip
inversions to finite element models. This program reads in the
triangle vertices from the program triangle (Shewchuck, 1996) or
a list and determines user-specified characteristics of the
network. Currently, it can write the outline of the mesh and the
neighboring elements to each triangle.