Add program interpolate.f90
This program is a wrapper for the subroutines in the new module
interpolation_module.f90. It reads the observation ("anchor")
points and desired interpolation points, and interpolates between
them, printing the estimated function value and its first and
second derivatives.

The interpolation module subroutines are enhanced to calculate the
first and second derivatives of the spline interpolation for the
new program.