Modify surf_disp.sh executable search, errors
Previously, the location of the Hdef binaries in surf_disp.sh was
defined as part of the installation process in the Makefile. To
generalize the search for executables, look in standard locations:
PATH environment variable, script directory, and ../bin (relative
to hdef-dir/scripts or hdef-dir/build). The user may also specify
HDEF_BIN_DIR on the command line.

Add error exits to script if programs exit with code!=0.

Add trap to remove temporary files even if script does not finish
executing.