Add slip threshold option to o92util
(This flag existed in a previous iteration of o92util, so I am not
entirely sure why it disappeared in the first place.) Add a flag
"-thr" to define a slip threshold for the input faults. Any faults
with slip less than the specified threshold have slip magnitudes
set to zero.

Update the o92util man page.

Correct the online help from "-autosta" to say "-auto".