Create contiguous arrays to avoid temporaries
Several subroutine calls in fltinv source and one pnpoly() call in
tri_disloc_module.f90 pass parts of arrays. These are often non-
contiguous blocks of the array, so a temporary array is made to
store the array in contiguous memory. This can have performance
implications, so these were replaced with temporary arrays stored
contiguously. Because of the nature of many of these arrays, the
performance improvement is limited.