Remove calls to usage() from read_halfspace_file()
Subroutine usage() was called in read_halfspace_file() from its
legacy position as a subroutine built into each program. This is
not a general design. To make more flexible, add argument ierr to
read_halfspace_file() and return non-zero value when an error
is encountered.

Update o92util, triutil, and fltinv calls to read_halfspace_file()
to call local usage statement when ierr is not zero.