Add option to resample anneal_post input ensemble
Because the models in the anneal log file might not necessarily be
distributed proportional to the posterior probability density
function, add the option to resample the available models based on
the objective function values saved in the log file. It is assumed
that these are log(probability) values.

The option -resample in anneal_post results in calling the new
subroutine resample() in the annealing_module, which performs the
calculation.

Update the Makefile template for anneal_post.