Move resampling in anneal_post to separate subroutine
Resampling originally required reading in all models and copying
the resampled models to temporary arrays. To save on memory for big
searches, read through the file twice: (1) read the log-probability
values and compute the resampled models; and (2) read the file
again, only saving models that are resampled.

This is more efficient if the resampled models are sorted, so
compile with sort module. Update Makefile template to include
sorting subroutines in anneal_post dependencies.