Adjust nit in anneal_post if set incorrectly in header
The number of iterations (nit) for the annealing log is set in
the header, but sometimes this number does not match the number of
iterations in the file. The header number should be the number of
models in the file minus 1 (since the initial model is 0).

When resampling, the header does not need to match exactly, since
there is no burn-in stage and models will be selected proportional
to their probability anyways. This commit updates nit during the
read so that the program will continue if the header is not
correct.