Add script for finding Hdef executable
Bash script takes an Hdef executable as an argument and looks for
it in the user's environment. The full path to the executable is
printed to hdefexec.tmp.

If a path (relative or absolute) is supplied, use that executable.
If no path is supplied, look for the executable in the directory
../build/ (assumes standard CMake directory structure), then in
the user's PATH.