Parallelize o92util at loop over stations
Most calculations of displacement, strain, or stress involve
multiple stations, so parallelizing the calculation over stations
can lead to better speedup than parallelizing over faults. Move
the OpenMP directives in o92util to define a parallel loop over
stations. The output still needs to be ordered, so enforce an
ordered parallel do loop. This can still go very fast if the
parallel schedule is predefined to have chunk sizes of 1 (i.e.,
the threads are fully interleaved). In fact, tests show that this
approach still runs (slightly) faster than parallelizing the fault
loop, even in cases with 1000s of faults and only 10 stations. I
suspect this is because the strain to stress calculation and
writing to files are now included in the parallel loop.