Calculate non-double couple scalar seismic moment
Existing subroutine pnt2mom() calculated the scalar seismic moment
for a double couple source. This is what GCMT used, so it was
working when I checked it. However, the USGS uses the more general
seismic moment defined in Equation 20 of Jost and Herrmann (1989).
This is also the scaling factor GMT psmeca uses. A new subroutine
pnt2mom_nondc() calculates the seismic moment for a general moment
tensor source.